# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Web app for code challenges with any environments",
        "title": "Goto",
        "contact": {
            "name": "Goto GitHub",
            "url": "http://github.com/shk3t/goto"
        }
    },
    "host": "localhost:3228",
    "basePath": "/api/",
    "paths": {
...
        "/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Проекты"
                ],
                "summary": "Список моих проектов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Вернуть с",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество возвращаемых элементов",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Язык",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название модуля",
                        "name": "module",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/goto_src_model.ProjectMin"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Проекты"
                ],
                "summary": "Добавление проекта с задачами",
                "parameters": [
                    {
                        "description": "Информация о проекте",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "type": "file",
                        "description": "Архив с проектом",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goto_src_model.DelayedTask"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Проекты"
                ],
                "summary": "Детализация моего проекта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор проекта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goto_src_model.ProjectPublic"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Проекты"
                ],
                "summary": "Обновление проекта с задачами",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор проекта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация о проекте",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "type": "file",
                        "description": "Архив с проектом",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goto_src_model.DelayedTask"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Проекты"
                ],
                "summary": "Удаление проекта с задачами",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "goto_src_model.DelayedTask": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "targetId": {
                    "type": "integer"
                },
                "targetName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "goto_src_model.ProjectMin": {
            "type": "object",
            "properties": {
                "containerization": {
                    "type": "string"
                },
                "failKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "srcDir": {
                    "type": "string"
                },
                "stubDir": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goto_src_model.TaskMin"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "goto_src_model.ProjectPublic": {
            "type": "object",
            "properties": {
                "containerization": {
                    "type": "string"
                },
                "failKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "srcDir": {
                    "type": "string"
                },
                "stubDir": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goto_src_model.Task"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "goto_src_model.Task": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goto_src_model.TaskFile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                },
                "runtarget": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "goto_src_model.TaskFile": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "stub": {
                    "type": "string"
                },
                "taskId": {
                    "type": "integer"
                }
            }
        },
...
        "goto_src_model.TaskMin": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Prepend your JWT key with `Bearer`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}